# swagger.yaml
openapi: 3.0.0
info:
  description: "API template"
  version: "0.0.0"
  title: "API template"
servers:
  - url: 'http://localhost:3000/'
paths:
  /meteo:
    get:
      tags: 
        - "meteo"
      parameters: 
        - in: query
          name: city
          description: Enter a city to get weather data
          required: true
          schema: 
            type: string
          example: Paris
        - in: query
          name: apiKey
          description: Enter your api key to have access to the open api weather
          required: true
          schema: 
            type: string
          example: dzad5z7da7dza5dad7r7e5zfz7f
      responses:
        '200':
          description: JSON Array of the city weather data
          content: 
            application/json:
              schema:
                type: array
                items:
                  type: string
        '400':
          description: City is missing. Required.
        '401':
          description: Unauthorized to get weather data. API key is missing or wrong api key.
        '404':
          description: City not found.
  /meteo/coord:
    get:
      tags:
        - "coord"
      parameters:
        - in: query
          name: lat
          description: Enter a latitude position
          schema:
            type: integer
          exemple: 25
        - in: query
          name: long
          description: Enter a longitude position
          schema:
            type: integer
          exemple: 20
        - in: query
          name: apiKey
          description: Enter your api key to have access to the open api weather
          required: true
          schema: 
            type: string
          example: dzad5z7da7dza5dad7r7e5zfz7f
      responses:
        '200':
          description: JSON array of the geographical coordinate weather data
          content:
            application/json:
              schema:
                type: array
                items: 
                  type: string
        '400':
          description: No data for those geographicals coordinates
        '401': 
          description: Unauthorized to get weather data. API key is missing or wrong api key.
          